[CLN] (test_)mail: cleanup controllers test setup / organization
It seems it can be cleanup a bit to better differentiate tests from
setup and tools. Guidelines in mail are: TestSomething are class
doing tests, while common have shorter name finishing with Common
indicating they own data and tools. Also it is better to keep a prefix
compared to a suffix as it is how tests are designed in (test_)mail,
hence using MailController suffix. Finally a tag is added, easing
finding controller tests, aka "mail_controller".

Renaming performed:
 * TestControllerCommon -> MailControllerCommon
 * TestAttachmentControllerCommon -> MailControllerAttachmentCommon
 * TestBinaryControllerCommon -> MailControllerBinaryCommon
 * TestMessageReactionControllerCommon -> MailControllerReactionCommon
 * TestMessageUpdateControllerCommon -> MailControllerUpdateCommon
 * TestThreadControllerCommon -> MailControllerThreadCommon

Rename files out of discuss zone to ease classification and coverage
understanding. Discuss files are globally left untouched as they are
out of SM management.

Move some tests from discuss subfolder to test_mail as they are
generic and should run on model agnostic records, not channel.

Add back some _name as it really helps grepping stuff.

Task-4332801 [test_mail] Cleanup controller tests
Prepares Task-4332797 : [mail] Partner from emails 3.0
Prepares Task-4273479 : [mail] Email-like recipients
