+++ addons/cloud_storage_azure/tests/test_cloud_storage_azure_attachment_controller.py
@@ -1,5 +1,3 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 import json
 import re
 from requests import Response
@@ -7,16 +5,12 @@ from unittest.mock import patch
 
 import odoo
 from odoo.tools.misc import file_open
-from odoo.addons.cloud_storage_azure.tests.test_cloud_storage_azure import (
-    TestCloudStorageAzureCommon,
-)
-from odoo.addons.mail.tests.test_attachment_controller import TestAttachmentControllerCommon
+from odoo.addons.cloud_storage_azure.tests.test_cloud_storage_azure import TestCloudStorageAzureCommon
+from odoo.addons.mail.tests.common_controllers import MailControllerAttachmentCommon
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestCloudStorageAttachmentController(
-    TestAttachmentControllerCommon, TestCloudStorageAzureCommon
-):
+@odoo.tests.tagged("-at_install", "post_install", "mail_controller")
+class TestCloudStorageAttachmentController(MailControllerAttachmentCommon, TestCloudStorageAzureCommon):
     def test_cloud_storage_azure_attachment_upload(self):
         """Test uploading an attachment with azure cloud storage."""
         thread = self.env["res.partner"].create({"name": "Test"})
