+++ addons/cloud_storage_google/tests/test_cloud_storage_google_attachment_controller.py
@@ -1,20 +1,14 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 import json
 import re
 
 import odoo
 from odoo.tools.misc import file_open
-from odoo.addons.cloud_storage_google.tests.test_cloud_storage_google import (
-    TestCloudStorageGoogleCommon,
-)
-from odoo.addons.mail.tests.test_attachment_controller import TestAttachmentControllerCommon
+from odoo.addons.cloud_storage_google.tests.test_cloud_storage_google import TestCloudStorageGoogleCommon
+from odoo.addons.mail.tests.common_controllers import MailControllerAttachmentCommon
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestCloudStorageAttachmentController(
-    TestAttachmentControllerCommon, TestCloudStorageGoogleCommon
-):
+@odoo.tests.tagged("-at_install", "post_install", "mail_controller")
+class TestCloudStorageAttachmentController(MailControllerAttachmentCommon, TestCloudStorageGoogleCommon):
     def test_cloud_storage_google_attachment_upload(self):
         """Test uploading an attachment with google cloud storage."""
         thread = self.env["res.partner"].create({"name": "Test"})
