+++ addons/mail/tests/common_tracking.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from collections import defaultdict
 from datetime import datetime, timedelta
 from unittest.mock import patch
