+++ addons/mail/tests/discuss/test_discuss_attachment_controller.py
@@ -1,11 +1,11 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import odoo
-from odoo.addons.mail.tests.test_attachment_controller import TestAttachmentControllerCommon
+from odoo.addons.mail.tests.common_controllers import MailControllerAttachmentCommon
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestDiscussAttachmentController(TestAttachmentControllerCommon):
+@odoo.tests.tagged("-at_install", "post_install", "mail_controller")
+class TestDiscussAttachmentController(MailControllerAttachmentCommon):
     def test_attachment_not_allowed_upload_public_channel(self):
         """Test access to upload an attachment on a not allowed upload public channel"""
         channel = self.env["discuss.channel"].create(
