+++ addons/mail/tests/discuss/test_discuss_binary_controller.py
@@ -1,11 +1,10 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-import odoo
-from odoo.addons.mail.tests.test_binary_controller import TestBinaryControllerCommon
+from odoo.addons.mail.tests.common_controllers import MailControllerBinaryCommon
+from odoo.tests import tagged
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestDiscussBinaryControllerCommon(TestBinaryControllerCommon):
+@tagged("-at_install", "post_install", "mail_controller")
+class TestDiscussBinaryController(MailControllerBinaryCommon):
+
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
@@ -19,9 +18,6 @@ class TestDiscussBinaryControllerCommon(TestBinaryControllerCommon):
             cls.user_public + cls.user_portal + cls.user_employee + cls.user_demo + cls.user_admin
         ).partner_id.ids
 
-
-@odoo.tests.tagged("-at_install", "post_install")
-class TestDiscussBinaryController(TestDiscussBinaryControllerCommon):
     def test_open_guest_avatar(self):
         """Test access to open the avatar of a guest.
         There is no common channel or any interaction from the guest."""
