+++ addons/project/tests/test_project_task_mail_tracking_duration.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.mail.tests.mail_tracking_duration_mixin_case import MailTrackingDurationMixinCase
+from odoo.addons.mail.tests.common_tracking import MailTrackingDurationMixinCase
 from odoo.tests import Form
