+++ addons/test_mail/security/ir.model.access.csv
@@ -57,6 +57,6 @@ access_mail_test_track_monetary,mail.test.track.monetary,model_mail_test_track_m
 access_mail_test_track_selection_portal,mail.test.track.selection.portal,model_mail_test_track_selection,base.group_portal,0,0,0,0
 access_mail_test_track_selection_user,mail.test.track.selection.user,model_mail_test_track_selection,base.group_user,1,1,1,1
 access_mail_test_properties_user,mail.test.properties.user,model_mail_test_properties,base.group_user,1,1,1,1
-access_mail_test_mail_tracking_duration,mail.test.mail.tracking.duration,model_mail_test_mail_tracking_duration,base.group_user,1,1,1,1
+access_mail_test_track_duration_mixin,mail.test.track.duration.mixin,model_mail_test_track_duration_mixin,base.group_user,1,1,1,1
 access_mail_test_public_user,mail.test.public.user,model_mail_test_public,base.group_user,1,1,1,1
 access_mail_test_public_public,mail.test.public.public,model_mail_test_public,base.group_public,1,1,0,0
