+++ addons/test_mail/tests/__init__.py
@@ -1,5 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import test_controller_binary
+from . import test_controller_thread
 from . import test_invite
 from . import test_ir_actions
 from . import test_ir_attachment
@@ -25,4 +27,3 @@ from . import test_message_management
 from . import test_message_post
 from . import test_message_track
 from . import test_performance
-from . import test_public_binary_controller
