+++ addons/test_mail/tests/test_controller_binary.py
@@ -1,11 +1,9 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-import odoo
-from odoo.addons.mail.tests.test_binary_controller import TestBinaryControllerCommon
+from odoo.addons.mail.tests.common_controllers import MailControllerBinaryCommon
+from odoo.tests import tagged
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestPublicBinaryController(TestBinaryControllerCommon):
+@tagged("-at_install", "post_install", "mail_controller")
+class TestPublicBinaryController(MailControllerBinaryCommon):
     def test_01_guest_avatar_public_record(self):
         """Test access to open a guest avatar who hasn't sent a message on a public record."""
         self.env["mail.test.public"].create({"name": "Test"})
