+++ addons/test_mail_full/tests/__init__.py
@@ -1,13 +1,13 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import test_odoobot
+from . import test_controller_attachment
+from . import test_controller_reaction
+from . import test_controller_update
+from . import test_controller_thread
+from . import test_mail_bot
 from . import test_mail_performance
 from . import test_mail_thread_internals
 from . import test_mass_mailing
-from . import test_message_reaction_controller_portal
 from . import test_portal
-from . import test_portal_attachment_controller
-from . import test_portal_message_update_controller
-from . import test_portal_thread_controller
 from . import test_rating
 from . import test_res_users
