+++ addons/test_mail_full/tests/test_controller_attachment.py
@@ -1,12 +1,12 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import odoo
-from odoo.addons.mail.tests.test_attachment_controller import TestAttachmentControllerCommon
+from odoo.addons.mail.tests.common_controllers import MailControllerAttachmentCommon
 from odoo.addons.portal.tests.test_portal_controller_common import TestPortalControllerCommon
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestPortalAttachmentController(TestAttachmentControllerCommon, TestPortalControllerCommon):
+@odoo.tests.tagged("-at_install", "post_install", "mail_controller")
+class TestPortalAttachmentController(MailControllerAttachmentCommon, TestPortalControllerCommon):
     def test_attachment_upload_portal(self):
         """Test access to upload an attachment on portal"""
         record = self.env["mail.test.portal.no.partner"].create({"name": "Test"})
