+++ addons/test_mail_full/tests/test_controller_reaction.py
@@ -1,16 +1,11 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-import odoo
-from odoo.addons.mail.tests.test_message_reaction_controller import (
-    TestMessageReactionControllerCommon,
-)
+from odoo.addons.mail.tests.common_controllers import MailControllerReactionCommon
 from odoo.addons.portal.tests.test_portal_controller_common import TestPortalControllerCommon
+from odoo.tests import tagged
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestPortalMessageReactionController(
-    TestMessageReactionControllerCommon, TestPortalControllerCommon
-):
+@tagged("-at_install", "post_install", "mail_controller")
+class TestPortalMessageReactionController(MailControllerReactionCommon, TestPortalControllerCommon):
+
     def test_message_reaction_portal_no_partner(self):
         """Test access of message reaction for portal without partner."""
         record = self.env["mail.test.portal.no.partner"].create({"name": "Test"})
