+++ addons/test_mail_full/tests/test_controller_thread.py
@@ -1,15 +1,10 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-import odoo
-from odoo.addons.mail.tests.test_thread_controller import (
-    MessagePostSubTestData,
-    TestThreadControllerCommon,
-)
+from odoo.addons.mail.tests.common_controllers import MailControllerThreadCommon, MessagePostSubTestData
 from odoo.addons.portal.tests.test_portal_controller_common import TestPortalControllerCommon
+from odoo.tests import tagged
 
 
-@odoo.tests.tagged("-at_install", "post_install")
-class TestPortalThreadController(TestThreadControllerCommon, TestPortalControllerCommon):
+@tagged("-at_install", "post_install", "mail_controller")
+class TestPortalThreadController(MailControllerThreadCommon, TestPortalControllerCommon):
     def test_message_post_access_portal_no_partner(self):
         """Test access of message post for portal without partner."""
         record = self.env["mail.test.portal.no.partner"].create({"name": "Test"})
